backend/requirements.txt
fastapi
uvicorn
Pillow
sqlalchemy
pymysql
python-dotenv